+++ CPP/Addidtion program in c.cpp
