+++ CPP/KMP Searching Algorithm in C++
