+++ Java/Duck_or_not.py
