+++ Java/Electricity bill calculator.java
