+++ Java/GuessGameDif.py
