+++ Java/Shuffle_array.py
