+++ Scripts/triangle.php
